pyproject.toml
[tool.poetry]
name = "devops"
version = "0.1.0"
description = ""
authors = ["Xavier Méndez Yánez <info@xavicode.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
diagram = "^0.2.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
